libseteec: Secure Element control
A secure element might be configured by OP-TEE; if the SE enabled
SCP03, the encryption keys should have been derived from the HUK and
never known to the normal world. In such circumstances, APDU frames
will need to be routed by the secure world as it is the only one with
access to the SCP03 keys.

By default, Secure Elements might be shipped with predefined keys
stored in persistent memory and documented in the device data
sheets.

This library provides an interface to enable SCP03 using those
non-secure keys. It also provides an interface to rotate these default
keys and derive board unique new ones.

A working scenario currently in OP-TEE is the NXP SE050 device.

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>